requirements.txt
numpy
pandas
tensorflow
keras
streamlit
Pillow
matplotlib
